+++ src/ui/src/js/controllers/admin_garden.js
@@ -311,6 +311,13 @@ export default function adminGardenController(
 
   }
 
+  $scope.hasConfiguredConnection = function(connections) {
+    if (connections !== undefined && connections != null){
+      return connections.some(connection => connection.status != 'NOT_CONFIGURED');
+    }
+    return false;
+  }
+
   $scope.eventUpsetGarden = function (garden) {
     if (garden.connection_type != "LOCAL" && !garden.has_parent) {
       for (let i = 0; i < $scope.data.length; i++) {